Better logs. Process invoke fixes.

+++ Managment/ReignOS.Core/ProcessUtil.cs
@@ -10,22 +10,42 @@ namespace ReignOS.Core
 {
     public static class ProcessUtil
     {
-        public static string Run(string name, string args, Dictionary<string,string>? enviromentVars = null, bool wait = true)
+        public static string Run(string name, string args, Dictionary<string,string>? enviromentVars = null, bool wait = true, bool asAdmin = false)
         {
             using (var process = new Process())
             {
-                process.StartInfo.FileName = name;
-                process.StartInfo.Arguments = args;
+                if (asAdmin)
+                {
+                    process.StartInfo.FileName = "sudo";
+                    process.StartInfo.Arguments = $"-S -- {name} {args}";
+                    process.StartInfo.RedirectStandardInput = true;
+                }
+                else
+                {
+                    process.StartInfo.FileName = name;
+                    process.StartInfo.Arguments = args;
+                }
                 if (enviromentVars != null)
                 {
                     foreach (var v in enviromentVars) process.StartInfo.EnvironmentVariables[v.Key] = v.Value;
                 }
                 process.StartInfo.RedirectStandardOutput = true;
+                process.StartInfo.RedirectStandardError = true;
                 process.Start();
+
+                if (asAdmin)
+                {
+                    process.StandardInput.WriteLine("gamer");
+                    process.StandardInput.Flush();
+                }
+
                 if (wait)
                 {
                     process.WaitForExit();
-                    return process.StandardOutput.ReadToEnd();
+                    var builder = new StringBuilder();
+                    builder.Append(process.StandardOutput.ReadToEnd());
+                    builder.Append(process.StandardError.ReadToEnd());
+                    return builder.ToString();
                 }
                 else
                 {
@@ -34,12 +54,19 @@ namespace ReignOS.Core
             }
         }
 
-        public static void Kill(string name)
+        public static void Kill(string name, bool asAdmin)
         {
-            foreach (var process in Process.GetProcessesByName(name))
+            if (asAdmin)
             {
-                process.Kill();
-                process.Dispose();
+                Run(name, "", null, wait:true, asAdmin:true);
+            }
+            else
+            {
+                foreach (var process in Process.GetProcessesByName(name))
+                {
+                    process.Kill();
+                    process.Dispose();
+                }
             }
         }
     }
